#[DEL] : 불필요 뷰 제거
